Newsfeeds Api and Tests
