+++ newsfeeds/admin.py
@@ -1,7 +1,7 @@
 from django.contrib import admin
-from newsfeeds.models import NewsFeeds
+from newsfeeds.models import NewsFeed
 
-@admin.register(NewsFeeds)
+@admin.register(NewsFeed)
 class NewsFeedAdmin(admin.ModelAdmin):
     list_display = ('user', 'tweet', 'created_at')
     date_hierarchy = 'created_at'
