+++ newsfeeds/api/serializers.py
@@ -1,0 +1,11 @@
+from rest_framework import serializers
+from newsfeeds.models import NewsFeed
+from tweets.api.serializers import TweetSerializer
+
+
+class NewsFeedSerializer(serializers.ModelSerializer):
+    tweet = TweetSerializer()
+
+    class Meta:
+        model = NewsFeed
+        fields = ('id', 'created_at', 'tweet')
