+++ testing/testcases.py
@@ -1,15 +1,23 @@
 from django.test import TestCase as DjangoTestCase
 from django.contrib.auth.models import User
 from tweets.models import Tweet
+from rest_framework.test import APIClient
 
 
 class TestCase(DjangoTestCase):
 
+    @property
+    def anonymous_client(self):
+        if hasattr(self, '_anonymous_client'):
+            return self._anonymous_client
+        self._anonymous_client = APIClient()
+        return self._anonymous_client
+
     def create_user(self, username, email=None, password=None):
+        if email is None:
+            email = '{}@jiuzhang.com'.format(username)
         if password is None:
             password = 'generic password'
-        if email is None:
-            email = f'{username}@twitter.com'
         # 不能写成 User.objects.create()
         # 因为 password 需要被加密, username 和 email 需要进行一些 normalize 处理
         return User.objects.create_user(username, email, password)
